+++ e3nn/o3/tensor_product.py
@@ -1,3 +1,4 @@
+from typing import Optional, List, Union
 from math import sqrt
 from collections import namedtuple
 import textwrap
@@ -15,8 +16,7 @@ def _prod(x):
     return out
 
 
-# This decorator applies to all the subclasses as well.
-@compile_mode('trace')
+@compile_mode('script')
 class TensorProduct(CodeGenMixin, torch.nn.Module):
     r"""Tensor Product with parametrizable paths
 
@@ -137,28 +137,39 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
     ...         if ir_out in ir_1 * ir_2
     ...     ]
     ... )
-    >>> with torch.no_grad():
-    ...     for weight in module.parameters():
-    ...         # formula from torch.nn.init.xavier_uniform_
+    >>> ws = []
+    >>> for ins in module.instructions:
+    ...     if ins.has_weight:
+    ...         weight = torch.empty(ins.path_shape)
     ...         mul_1, mul_2, mul_out = weight.shape
+    ...         # formula from torch.nn.init.xavier_uniform_
     ...         a = (6 / (mul_1 * mul_2 + mul_out))**0.5
-    ...         _ = weight.uniform_(-a, a)  # `_ = ` is only here because of pytest
+    ...         ws += [weight.uniform_(-a, a).view(-1)]
+    >>> with torch.no_grad():
+    ...     module.weight[:] = torch.cat(ws)
     >>> n = 1_000
     >>> vars = module(irreps_1.randn(n, -1), irreps_2.randn(n, -1)).var(0)
     >>> assert vars.min() > 1 / 3
     >>> assert vars.max() < 3
     """
+
+    _profiling_str: str
+    shared_weights: bool
+    internal_weights: bool
+    weight_numel: int
+
     def __init__(
-            self,
-            in1,
-            in2,
-            out,
-            instructions,
-            normalization='component',
-            internal_weights=None,
-            shared_weights=None,
-            _specialized_code=True,
-                ):
+        self,
+        in1,
+        in2,
+        out,
+        instructions,
+        normalization='component',
+        internal_weights=None,
+        shared_weights=None,
+        _specialized_code=True,
+    ):
+        # === Setup ===
         super().__init__()
 
         assert normalization in ['component', 'norm'], normalization
@@ -199,27 +210,57 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
         in2_var = [var for _, _, var in in2]
         out_var = [var for _, _, var in out]
 
+        # === Build instructions ===
+        Instruction = namedtuple("Instruction", "i_in1, i_in2, i_out, connection_mode, has_weight, path_weight, path_shape")
+        instructions = [x if len(x) == 6 else x + (1.0,) for x in instructions]
+        self.instructions = [
+            Instruction(
+                i_in1, i_in2, i_out, connection_mode, has_weight, path_weight,
+                {
+                    'uvw': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul, self.irreps_out[i_out].mul),
+                    'uvu': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul),
+                    'uvv': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul),
+                    'uuw': (self.irreps_in1[i_in1].mul, self.irreps_out[i_out].mul),
+                    'uuu': (self.irreps_in1[i_in1].mul,),
+                    'uvuv': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul),
+                }[connection_mode],
+            )
+            for i_in1, i_in2, i_out, connection_mode, has_weight, path_weight in instructions
+        ]
+
+        # === Determine weights ===
         self.shared_weights = shared_weights
-        z = '' if self.shared_weights else 'z'
+        self.weight_numel = sum(_prod(ins.path_shape) for ins in self.instructions if ins.has_weight)
 
+        self.internal_weights = internal_weights
+        if internal_weights and self.weight_numel > 0:
+            assert self.shared_weights, "Having internal weights impose shared weights"
+            self.weight = torch.nn.Parameter(torch.randn(self.weight_numel))
+        else:
+            # For TorchScript, there always has to be some kind of defined .weight
+            self.register_buffer('weight', torch.Tensor())
+
+        # === Init everything for codegen ===
         # == TorchScript main operation templates ==
-        # The if-else block is needed to avoid an internal TorchScript compiler bug related to the early return.
-        code_out = textwrap.dedent(f"""
+        code_header = textwrap.dedent("""
         from typing import List
-
         import torch
-
         from e3nn.util import broadcast_tensors
-
+        """)
+        # The if-else block is needed to avoid an internal TorchScript compiler bug related to the early return.
+        funcdef_out = textwrap.dedent(f"""
         @torch.jit.script
-        def main(x1: torch.Tensor, x2: torch.Tensor, ws: List[torch.Tensor], w3j: List[torch.Tensor]) -> torch.Tensor:
-            x1, x2 = broadcast_tensors(x1, x2)
+        def main(x1: torch.Tensor, x2: torch.Tensor, ws: torch.Tensor, w3j: torch.Tensor) -> torch.Tensor:
+            {'x1, x2 = broadcast_tensors(x1, x2)' if shared_weights else ''}
+            {'x1, x2, ws = torch.broadcast_tensors(x1[..., :, None, None], x2[..., None, :, None], ws[..., None, None, :])' if not shared_weights else ''}
+            {'x1, x2, ws = x1[..., :, 0, 0], x2[..., 0, :, 0], ws[..., 0, 0, :]' if not shared_weights else ''}
             size = x1.shape[:-1]
             outsize = size + ({self.irreps_out.dim},)
             assert x1.shape[-1] == {self.irreps_in1.dim}, "Incorrect feature dimension for x1"
             assert x2.shape[-1] == {self.irreps_in2.dim}, "Incorrect feature dimension for x2"
             x1 = x1.reshape(-1, {self.irreps_in1.dim})
             x2 = x2.reshape(-1, {self.irreps_in2.dim})
+            ws = ws.reshape(-1, {self.weight_numel})
 
             if x1.shape[0] == 0:
                 return x1.new_zeros(outsize)
@@ -229,19 +270,15 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                 ein = torch.einsum
         """)
 
-        code_right = textwrap.dedent(f"""
-        from typing import List
-
-        import torch
-
-        from e3nn.util import broadcast_tensors
-
+        funcdef_right = textwrap.dedent(f"""
         @torch.jit.script
-        def main(x2: torch.Tensor, ws: List[torch.Tensor], w3j: List[torch.Tensor]) -> torch.Tensor:
+        def main(x2: torch.Tensor, ws: torch.Tensor, w3j: torch.Tensor) -> torch.Tensor:
+            {'x2, ws = broadcast_tensors(x2, ws)' if not shared_weights else ''}
             size = x2.shape[:-1]
             outsize = size + ({self.irreps_in1.dim}, {self.irreps_out.dim},)
             assert x2.shape[-1] == {self.irreps_in2.dim}, "Incorrect feature dimension for x2"
             x2 = x2.reshape(-1, {self.irreps_in2.dim})
+            ws = ws.reshape(-1, {self.weight_numel})
 
             if x2.shape[0] == 0:
                 return x2.new_zeros(outsize)
@@ -252,6 +289,8 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
         """)
         # == end TorchScript templates ==
         # Put everything in the else block
+        code_out = ""
+        code_right = ""
         base_indent = 2
         def indent_for_level(indent_level):
             return ((base_indent + indent_level) * 4) * " "
@@ -274,26 +313,11 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
         code_out += "\n"
         code_right += "\n"
 
+        z = '' if self.shared_weights else 'z'
         last_ss = None
 
-        Instruction = namedtuple("Instruction", "i_in1, i_in2, i_out, connection_mode, has_weight, path_weight, weight_shape")
-        instructions = [x if len(x) == 6 else x + (1.0,) for x in instructions]
-        self.instructions = [
-            Instruction(
-                i_in1, i_in2, i_out, connection_mode, has_weight, path_weight,
-                {
-                    'uvw': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul, self.irreps_out[i_out].mul),
-                    'uvu': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul),
-                    'uvv': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul),
-                    'uuw': (self.irreps_in1[i_in1].mul, self.irreps_out[i_out].mul),
-                    'uuu': (self.irreps_in1[i_in1].mul,),
-                    'uvuv': (self.irreps_in1[i_in1].mul, self.irreps_in2[i_in2].mul),
-                }[connection_mode] if has_weight else None
-            )
-            for i_in1, i_in2, i_out, connection_mode, has_weight, path_weight in instructions
-        ]
-
         index_w = -1
+        flat_weight_i = 0
 
         for ins in self.instructions:
             mul_1, (l_1, p_1) = self.irreps_in1[ins.i_in1]
@@ -346,6 +370,11 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
 
             if ins.has_weight:
                 index_w += 1
+                # Extract the weight from the flattened weight tensor
+                line = f"{s}ws_{index_w} = ws[:, {flat_weight_i}:{flat_weight_i + _prod(ins.path_shape)}].reshape({(() if self.shared_weights else (-1,)) + tuple(ins.path_shape)})\n"
+                code_out += line
+                code_right += line
+                flat_weight_i += _prod(ins.path_shape)
 
             line_out = f"{s}out[:, {index_out}:{index_out+dim_out}] += {alpha} * {{}}.reshape(batch, {dim_out})\n\n"
             line_right = f"{s}out[:, {index_1}:{index_1+dim_1}, {index_out}:{index_out+dim_out}] += {alpha} * {{}}.reshape(batch, {dim_1}, {dim_out})\n\n"
@@ -365,11 +394,11 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                     code_right += line
 
                     if ins.connection_mode == 'uvw':
-                        code_out += line_out.format(f"ein('{z}uvw,zu,zv->zw', ws[{index_w}], s1, s2)")
-                        code_right += line_right.format(f"ein('{z}uvw,zv->zuw', ws[{index_w}], s2)")
+                        code_out += line_out.format(f"ein('{z}uvw,zu,zv->zw', ws_{index_w}, s1, s2)")
+                        code_right += line_right.format(f"ein('{z}uvw,zv->zuw', ws_{index_w}, s2)")
                     if ins.connection_mode == 'uvu':
-                        code_out += line_out.format(f"ein('{z}uv,zu,zv->zu', ws[{index_w}], s1, s2)")
-                        code_right += line_right.format(f"ein('{z}uv,uw,zv->zuw', ws[{index_w}], e1, s2)")
+                        code_out += line_out.format(f"ein('{z}uv,zu,zv->zu', ws_{index_w}, s1, s2)")
+                        code_right += line_right.format(f"ein('{z}uv,uw,zv->zuw', ws_{index_w}, e1, s2)")
 
                     continue
 
@@ -377,11 +406,11 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                     code_out += f"{s}s1 = s1.reshape(batch, {mul_1})\n"
 
                     if ins.connection_mode == 'uvw':
-                        code_out += line_out.format(f"ein('{z}uvw,zu,zvi->zwi', ws[{index_w}], s1, s2)")
-                        code_right += line_right.format(f"ein('{z}uvw,zvi->zuwi', ws[{index_w}], s2)")
+                        code_out += line_out.format(f"ein('{z}uvw,zu,zvi->zwi', ws_{index_w}, s1, s2)")
+                        code_right += line_right.format(f"ein('{z}uvw,zvi->zuwi', ws_{index_w}, s2)")
                     if ins.connection_mode == 'uvu':
-                        code_out += line_out.format(f"ein('{z}uv,zu,zvi->zui', ws[{index_w}], s1, s2)")
-                        code_right += line_right.format(f"ein('{z}uv,uw,zvi->zuwi', ws[{index_w}], e1, s2)")
+                        code_out += line_out.format(f"ein('{z}uv,zu,zvi->zui', ws_{index_w}, s1, s2)")
+                        code_right += line_right.format(f"ein('{z}uv,uw,zvi->zuwi', ws_{index_w}, e1, s2)")
 
                     continue
 
@@ -391,30 +420,30 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                     code_right += f"{s}wig = torch.eye({2 * l_1 + 1}, dtype=x2.dtype, device=x2.device)\n"
 
                     if ins.connection_mode == 'uvw':
-                        code_out += line_out.format(f"ein('{z}uvw,zui,zv->zwi', ws[{index_w}], s1, s2)")
-                        code_right += line_right.format(f"ein('{z}uvw,ij,zv->zuiwj', ws[{index_w}], wig, s2)")
+                        code_out += line_out.format(f"ein('{z}uvw,zui,zv->zwi', ws_{index_w}, s1, s2)")
+                        code_right += line_right.format(f"ein('{z}uvw,ij,zv->zuiwj', ws_{index_w}, wig, s2)")
                     if ins.connection_mode == 'uvu':
-                        code_out += line_out.format(f"ein('{z}uv,zui,zv->zui', ws[{index_w}], s1, s2)")
-                        code_right += line_right.format(f"ein('{z}uv,ij,uw,zv->zuiwj', ws[{index_w}], wig, e1, s2)")
+                        code_out += line_out.format(f"ein('{z}uv,zui,zv->zui', ws_{index_w}, s1, s2)")
+                        code_right += line_right.format(f"ein('{z}uv,ij,uw,zv->zuiwj', ws_{index_w}, wig, e1, s2)")
 
                     continue
 
                 if l_1 == l_2 and l_out == 0 and ins.connection_mode == 'uvw' and normalization == 'component' and ins.has_weight:
                     # Cl_l_0 = eye / sqrt(2L+1)
-                    code_out += line_out.format(f"ein('{z}uvw,zui,zvi->zw', ws[{index_w}] / {sqrt(2 * l_1 + 1)}, s1, s2)")
-                    code_right += line_right.format(f"ein('{z}uvw,zvi->zuiw', ws[{index_w}] / {sqrt(2 * l_1 + 1)}, s2)")
+                    code_out += line_out.format(f"ein('{z}uvw,zui,zvi->zw', ws_{index_w} / {sqrt(2 * l_1 + 1)}, s1, s2)")
+                    code_right += line_right.format(f"ein('{z}uvw,zvi->zuiw', ws_{index_w} / {sqrt(2 * l_1 + 1)}, s2)")
                     continue
 
                 if l_1 == l_2 and l_out == 0 and ins.connection_mode == 'uvu' and normalization == 'component' and ins.has_weight:
                     # Cl_l_0 = eye / sqrt(2L+1)
-                    code_out += line_out.format(f"ein('{z}uv,zui,zvi->zu', ws[{index_w}] / {sqrt(2 * l_1 + 1)}, s1, s2)")
-                    code_right += line_right.format(f"ein('{z}uv,uw,zvi->zuiw', ws[{index_w}] / {sqrt(2 * l_1 + 1)}, e1, s2)")
+                    code_out += line_out.format(f"ein('{z}uv,zui,zvi->zu', ws_{index_w} / {sqrt(2 * l_1 + 1)}, s1, s2)")
+                    code_right += line_right.format(f"ein('{z}uv,uw,zvi->zuiw', ws_{index_w} / {sqrt(2 * l_1 + 1)}, e1, s2)")
                     continue
 
                 if l_1 == l_2 and l_out == 0 and ins.connection_mode == 'uuu' and normalization == 'component' and ins.has_weight:
                     # Cl_l_0 = eye / sqrt(2L+1)
-                    code_out += line_out.format(f"ein('{z}u,zui,zui->zu', ws[{index_w}] / {sqrt(2 * l_1 + 1)}, s1, s2)")
-                    code_right += line_right.format(f"ein('{z}u,uw,zui->zuiw', ws[{index_w}] / {sqrt(2 * l_1 + 1)}, e1, s2)")
+                    code_out += line_out.format(f"ein('{z}u,zui,zui->zu', ws_{index_w} / {sqrt(2 * l_1 + 1)}, s1, s2)")
+                    code_right += line_right.format(f"ein('{z}u,uw,zui->zuiw', ws_{index_w} / {sqrt(2 * l_1 + 1)}, e1, s2)")
                     continue
 
                 if l_1 == l_2 and l_out == 0 and ins.connection_mode == 'uuu' and normalization == 'component' and not ins.has_weight:
@@ -428,7 +457,7 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                     code_out += f"{s}s1 = s1.reshape(batch, {mul_1}, 1, {2 * l_1 + 1})\n"
                     code_out += f"{s}s2 = s2.reshape(batch, 1, {mul_2}, {2 * l_2 + 1})\n"
                     code_out += f"{s}s1, s2 = torch.broadcast_tensors(s1, s2)\n"
-                    code_out += line_out.format(f"ein('{z}uvw,zuvi->zwi', ws[{index_w}] / {sqrt(2)}, torch.cross(s1, s2, dim=3))")
+                    code_out += line_out.format(f"ein('{z}uvw,zuvi->zwi', ws_{index_w} / {sqrt(2)}, torch.cross(s1, s2, dim=3))")
 
                     if (l_1, l_2, l_out) in wigners:
                         index_w3j = wigners.index((l_1, l_2, l_out))
@@ -436,7 +465,7 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                         index_w3j = len(wigners)
                         wigners += [(l_1, l_2, l_out)]
 
-                    code_right += line_right.format(f"ein('{z}uvw,ijk,zvj->zuiwk', ws[{index_w}], w3j[{index_w3j}], s2)")
+                    code_right += line_right.format(f"ein('{z}uvw,ijk,zvj->zuiwk', ws_{index_w}, w3j_{index_w3j}, s2)")
                     continue
 
                 if (l_1, l_2, l_out) == (1, 1, 1) and ins.connection_mode == 'uvu' and normalization == 'component' and ins.has_weight:
@@ -444,7 +473,7 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                     code_out += f"{s}s1 = s1.reshape(batch, {mul_1}, 1, {2 * l_1 + 1})\n"
                     code_out += f"{s}s2 = s2.reshape(batch, 1, {mul_2}, {2 * l_2 + 1})\n"
                     code_out += f"{s}s1, s2 = torch.broadcast_tensors(s1, s2)\n"
-                    code_out += line_out.format(f"ein('{z}uv,zuvi->zui', ws[{index_w}] / {sqrt(2)}, torch.cross(s1, s2, dim=3))")
+                    code_out += line_out.format(f"ein('{z}uv,zuvi->zui', ws_{index_w} / {sqrt(2)}, torch.cross(s1, s2, dim=3))")
 
                     if (l_1, l_2, l_out) in wigners:
                         index_w3j = wigners.index((l_1, l_2, l_out))
@@ -452,8 +481,9 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                         index_w3j = len(wigners)
                         wigners += [(l_1, l_2, l_out)]
 
-                    code_right += line_right.format(f"ein('{z}uv,ijk,uw,zvj->zuiwk', ws[{index_w}], w3j[{index_w3j}], e1, s2)")
+                    code_right += line_right.format(f"ein('{z}uv,ijk,uw,zvj->zuiwk', ws_{index_w}, w3j_{index_w3j}, e1, s2)")
                     continue
+            # == end specialized code ==
 
             if last_ss != (ins.i_in1, ins.i_in2, ins.connection_mode[:2]):
                 if ins.connection_mode[:2] == 'uv':
@@ -470,64 +500,83 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
 
             if ins.connection_mode == 'uvw':
                 assert ins.has_weight
-                code_out += line_out.format(f"ein('{z}uvw,ijk,zuvij->zwk', ws[{index_w}], w3j[{index_w3j}], ss)")
-                code_right += line_right.format(f"ein('{z}uvw,ijk,zvj->zuiwk', ws[{index_w}], w3j[{index_w3j}], s2)")
+                code_out += line_out.format(f"ein('{z}uvw,ijk,zuvij->zwk', ws_{index_w}, w3j_{index_w3j}, ss)")
+                code_right += line_right.format(f"ein('{z}uvw,ijk,zvj->zuiwk', ws_{index_w}, w3j_{index_w3j}, s2)")
             if ins.connection_mode == 'uvu':
                 assert mul_1 == mul_out
                 if ins.has_weight:
-                    code_out += line_out.format(f"ein('{z}uv,ijk,zuvij->zuk', ws[{index_w}], w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('{z}uv,ijk,uw,zvj->zuiwk', ws[{index_w}], w3j[{index_w3j}], e1, s2)")
+                    code_out += line_out.format(f"ein('{z}uv,ijk,zuvij->zuk', ws_{index_w}, w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('{z}uv,ijk,uw,zvj->zuiwk', ws_{index_w}, w3j_{index_w3j}, e1, s2)")
                 else:
-                    code_out += line_out.format(f"ein('ijk,zuvij->zuk', w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('ijk,uw,zvj->zuiwk', w3j[{index_w3j}], e1, s2)")
+                    code_out += line_out.format(f"ein('ijk,zuvij->zuk', w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('ijk,uw,zvj->zuiwk', w3j_{index_w3j}, e1, s2)")
             if ins.connection_mode == 'uvv':
                 assert mul_2 == mul_out
                 if ins.has_weight:
-                    code_out += line_out.format(f"ein('{z}uv,ijk,zuvij->zvk', ws[{index_w}], w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('{z}uv,ijk,zvj->zuivk', ws[{index_w}], w3j[{index_w3j}], s2)")
+                    code_out += line_out.format(f"ein('{z}uv,ijk,zuvij->zvk', ws_{index_w}, w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('{z}uv,ijk,zvj->zuivk', ws_{index_w}, w3j_{index_w3j}, s2)")
                 else:
-                    code_out += line_out.format(f"ein('ijk,zuvij->zvk', w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('u,ijk,zvj->zuivk', s2.new_zeros({mul_1}).fill_(1.0), w3j[{index_w3j}], s2)")
+                    code_out += line_out.format(f"ein('ijk,zuvij->zvk', w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('u,ijk,zvj->zuivk', s2.new_zeros({mul_1}).fill_(1.0), w3j_{index_w3j}, s2)")
             if ins.connection_mode == 'uuw':
                 assert mul_1 == mul_2
                 if ins.has_weight:
-                    code_out += line_out.format(f"ein('{z}uw,ijk,zuij->zwk', ws[{index_w}], w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('{z}uw,ijk,zuj->zuiwk', ws[{index_w}], w3j[{index_w3j}], s2)")
+                    code_out += line_out.format(f"ein('{z}uw,ijk,zuij->zwk', ws_{index_w}, w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('{z}uw,ijk,zuj->zuiwk', ws_{index_w}, w3j_{index_w3j}, s2)")
                 else:
                     assert mul_out == 1
-                    code_out += line_out.format(f"ein('ijk,zuij->zk', w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('ijk,zuj->zuik', w3j[{index_w3j}], s2)")
+                    code_out += line_out.format(f"ein('ijk,zuij->zk', w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('ijk,zuj->zuik', w3j_{index_w3j}, s2)")
             if ins.connection_mode == 'uuu':
                 assert mul_1 == mul_2 == mul_out
                 if ins.has_weight:
-                    code_out += line_out.format(f"ein('{z}u,ijk,zuij->zuk', ws[{index_w}], w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('{z}u,ijk,uw,zuj->zuiwk', ws[{index_w}], w3j[{index_w3j}], e1, s2)")
+                    code_out += line_out.format(f"ein('{z}u,ijk,zuij->zuk', ws_{index_w}, w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('{z}u,ijk,uw,zuj->zuiwk', ws_{index_w}, w3j_{index_w3j}, e1, s2)")
                 else:
-                    code_out += line_out.format(f"ein('ijk,zuij->zuk', w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('ijk,uw,zuj->zuiwk', w3j[{index_w3j}], e1, s2)")
+                    code_out += line_out.format(f"ein('ijk,zuij->zuk', w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('ijk,uw,zuj->zuiwk', w3j_{index_w3j}, e1, s2)")
             if ins.connection_mode == 'uvuv':
                 assert mul_1 * mul_2 == mul_out
                 if ins.has_weight:
-                    code_out += line_out.format(f"ein('{z}uv,ijk,zuvij->zuvk', ws[{index_w}], w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('{z}uv,ijk,uw,zvj->zuiwvk', ws[{index_w}], w3j[{index_w3j}], e1, s2)")
+                    code_out += line_out.format(f"ein('{z}uv,ijk,zuvij->zuvk', ws_{index_w}, w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('{z}uv,ijk,uw,zvj->zuiwvk', ws_{index_w}, w3j_{index_w3j}, e1, s2)")
                 else:
-                    code_out += line_out.format(f"ein('ijk,zuvij->zuvk', w3j[{index_w3j}], ss)")
-                    code_right += line_right.format(f"ein('ijk,uw,zvj->zuiwvk', w3j[{index_w3j}], e1, s2)")
+                    code_out += line_out.format(f"ein('ijk,zuvij->zuvk', w3j_{index_w3j}, ss)")
+                    code_right += line_right.format(f"ein('ijk,uw,zvj->zuiwvk', w3j_{index_w3j}, e1, s2)")
             code_out += "\n"
 
         code_out += f"{s}return out.reshape(outsize)"
         code_right += f"{s}return out.reshape(outsize)"
 
+        # w3j
+        wigner_mats = []
+        flat_wigner_index = 0
+        code_wigners = []
+        s = indent_for_level(0)
+        for i, (l_1, l_2, l_out) in enumerate(wigners):
+            wig = o3.wigner_3j(l_1, l_2, l_out)
+
+            if normalization == 'component':
+                wig *= (2 * l_out + 1) ** 0.5
+            if normalization == 'norm':
+                wig *= (2 * l_1 + 1) ** 0.5 * (2 * l_2 + 1) ** 0.5
+
+            code_wigners.append(f"{s}w3j_{i} = w3j[{flat_wigner_index}:{flat_wigner_index + _prod(wig.shape)}].reshape({tuple(wig.shape)})")
+            flat_wigner_index += _prod(wig.shape)
+            wigner_mats.append(wig)
+        code_wigners = '\n'.join(code_wigners)
+
+        if len(wigner_mats) > 0:
+            self.register_buffer('_wigner_buf', torch.cat([w.reshape(-1) for w in wigner_mats]))
+        else:
+            # We register an empty buffer so that call signatures don't have to change
+            self.register_buffer('_wigner_buf', torch.Tensor())
+
+        # Finalize the code
         if self.irreps_in1.dim == 0 or self.irreps_in2.dim == 0 or self.irreps_out.dim == 0:
-            code_out = textwrap.dedent(f"""
-            from typing import List
-
-            import torch
-
-            from e3nn.util import broadcast_tensors
-
+            full_code_out = code_header + textwrap.dedent(f"""
             @torch.jit.script
-            def main(x1: torch.Tensor, x2: torch.Tensor, ws: List[torch.Tensor], w3j: List[torch.Tensor]) -> torch.Tensor:
+            def main(x1: torch.Tensor, x2: torch.Tensor, ws: torch.Tensor, w3j: torch.Tensor) -> torch.Tensor:
                 x1, x2 = broadcast_tensors(x1, x2)
                 size = x1.shape[:-1]
                 outsize = size + ({self.irreps_out.dim},)
@@ -537,50 +586,24 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
                 return x1.new_zeros(outsize)
             """)
 
-            code_right = textwrap.dedent(f"""
-            from typing import List
-
-            import torch
-
-            from e3nn.util import broadcast_tensors
-
+            full_code_right = code_header + textwrap.dedent(f"""
             @torch.jit.script
-            def main(x2: torch.Tensor, ws: List[torch.Tensor], w3j: List[torch.Tensor]) -> torch.Tensor:
+            def main(x2: torch.Tensor, ws: torch.Tensor, w3j: torch.Tensor) -> torch.Tensor:
                 size = x2.shape[:-1]
                 outsize = size + ({self.irreps_in1.dim}, {self.irreps_out.dim},)
                 assert x2.shape[-1] == {self.irreps_in2.dim}, "Incorrect feature dimension for x2"
 
                 return x2.new_zeros(outsize)
             """)
+        else:
+            # paste together the pieces
+            full_code_out = "\n\n".join([code_header, funcdef_out, code_wigners, code_out])
+            full_code_right = "\n\n".join([code_header, funcdef_right, code_wigners, code_right])
 
         self._codegen_register({
-            '_compiled_main_out': code_out,
-            '_compiled_main_right': code_right,
+            '_compiled_main_out': full_code_out,
+            '_compiled_main_right': full_code_right,
         })
-
-        # w3j
-        self.wigners = wigners
-        for i, (l_1, l_2, l_out) in enumerate(self.wigners):
-            wig = o3.wigner_3j(l_1, l_2, l_out)
-
-            if normalization == 'component':
-                wig *= (2 * l_out + 1) ** 0.5
-            if normalization == 'norm':
-                wig *= (2 * l_1 + 1) ** 0.5 * (2 * l_2 + 1) ** 0.5
-
-            self.register_buffer(f"C{i}", wig)
-
-        # weights
-        self.weight_numel = sum(_prod(ins.weight_shape) for ins in self.instructions if ins.has_weight)
-
-        self.internal_weights = internal_weights
-        if internal_weights:
-            assert self.shared_weights, "Having internal weights impose shared weights"
-            self.weight = torch.nn.ParameterDict()
-            for ins in self.instructions:
-                if ins.has_weight:
-                    name = f'[{ins.i_in1}:{self.irreps_in1[ins.i_in1]}] x [{ins.i_in2}:{self.irreps_in2[ins.i_in2]}] -> [{ins.i_out}:{self.irreps_out[ins.i_out]}]'
-                    self.weight[name] = torch.nn.Parameter(torch.randn(ins.weight_shape))
 
         if self.irreps_out.dim > 0:
             output_mask = torch.cat([
@@ -593,116 +616,48 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
             output_mask = torch.ones(0)
         self.register_buffer('output_mask', output_mask)
 
+        # For TorchScript, this needs to be done in advance:
+        self._profiling_str = str(self)
+
     def __repr__(self):
-        npath = sum(
-            {
-                'uvw': self.irreps_in1[i.i_in1].mul * self.irreps_in2[i.i_in2].mul * self.irreps_out[i.i_out].mul,
-                'uvu': self.irreps_in1[i.i_in1].mul * self.irreps_in2[i.i_in2].mul,
-                'uvv': self.irreps_in1[i.i_in1].mul * self.irreps_in2[i.i_in2].mul,
-                'uuw': self.irreps_in1[i.i_in1].mul * self.irreps_out[i.i_out].mul,
-                'uuu': self.irreps_in1[i.i_in1].mul,
-                'uvuv': self.irreps_in1[i.i_in1].mul * self.irreps_in2[i.i_in2].mul,
-            }[i.connection_mode]
-            for i in self.instructions
-        )
+        npath = sum(_prod(i.path_shape) for i in self.instructions)
         return (
             f"{self.__class__.__name__}"
             f"({self.irreps_in1.simplify()} x {self.irreps_in2.simplify()} "
             f"-> {self.irreps_out.simplify()} | {npath} paths | {self.weight_numel} weights)"
         )
 
-    def prepare_weight_list(self, weight):
-        if self.weight_numel:
-            weight_shapes = [ins.weight_shape for ins in self.instructions if ins.has_weight]
-
-            if weight is None:
-                if not self.internal_weights:
-                    raise RuntimeError("Weights must be provided when the TensorProduct does not have `internal_weights`")
-                weight = list(self.weight.values())
-            if torch.is_tensor(weight):
-                ws = []
-                i = 0
-                for shape in weight_shapes:
-                    d = _prod(shape)
-                    if not self.shared_weights:
-                        ws += [weight[..., i:i+d].reshape((-1,) + shape)]
-                    else:
-                        ws += [weight[i:i+d].reshape(shape)]
-                    i += d
-                weight = ws
-            if isinstance(weight, list):
-                if not self.shared_weights:
-                    weight = [w.reshape(-1, *shape) for w, shape in zip(weight, weight_shapes)]
-                else:
-                    weight = [w.reshape(*shape) for w, shape in zip(weight, weight_shapes)]
-        else:
-            weight = []
-        return weight
-
-    def _make_tracing_inputs(self, n: int):
-        import inspect
-        import random
-        MAX_N_BATCH = 3
-        # - Inspect forward() -
-        # Check if forward has weights:
-        forward_params = list(inspect.signature(self.forward).parameters)
-        arg_generators = []
-        forward_has_weight = forward_params[-1] == 'weight'
-        if forward_has_weight:
-            forward_params = forward_params[:-1]
-
-        if len(forward_params) == 1:
-            # Something like Norm or Linear with irreps_in
-            arg_generators.append(
-                lambda bdim: self.irreps_in.randn(bdim, -1)
-            )
-        elif len(forward_params) == 2:
-            # A normal TensorProduct with two inputs
-            arg_generators.append(
-                lambda bdim: self.irreps_in1.randn(bdim, -1)
-            )
-            arg_generators.append(
-                lambda bdim: self.irreps_in2.randn(bdim, -1)
-            )
-        else:
-            raise NotImplementedError("Tried to trace compile an atypical TensorProduct subclass; please override make_tracing_input() for your class.")
-
-        # - Deal with weights -
-        # We only trace a weight input if the TP has no internal weights:
-        # in TorchScript, a TP either always uses internal weights or always
-        # uses input weights
-        trace_weights = bool(self.weight_numel) and not self.internal_weights
-        # No reason to trace weights when there aren't any, even if they aren't internal:
-        trace_weights = trace_weights and self.weight_numel > 0
-        make_weights = None
-        if trace_weights:
-            if self.shared_weights:
-                def make_weights(bdim):
-                    return torch.randn(self.weight_numel)
+    @torch.jit.unused
+    def _prep_weights_python(self, weight: Optional[Union[torch.Tensor, List[torch.Tensor]]]) -> Optional[torch.Tensor]:
+        if isinstance(weight, list):
+            weight_shapes = [ins.path_shape for ins in self.instructions if ins.has_weight]
+            if not self.shared_weights:
+                weight = [w.reshape(-1, _prod(shape)) for w, shape in zip(weight, weight_shapes)]
             else:
-                # By tracing it this way, we do enforce that non-shared weights are always passed in (..., self.weight_numel)
-                # form, which is fine, since lists don't play nice with compilation anyway
-                def make_weights(bdim):
-                    return torch.randn(bdim, self.weight_numel)
-            arg_generators.append(make_weights)
+                weight = [w.reshape(_prod(shape)) for w, shape in zip(weight, weight_shapes)]
+            return torch.cat(weight, dim=-1)
+        else:
+            return weight
 
-        # - Make outputs -
-        out = []
-        for _ in range(n):
-            # Choose a random batch dimension
-            bdim = random.randint(1, MAX_N_BATCH)
-            # right() always takes irreps_in2 so we build it here:
-            right_args = (self.irreps_in2.randn(bdim, -1),)
-            if trace_weights:
-                right_args = right_args + (make_weights(bdim),)
-            # put it in
-            out.append({
-                'forward': tuple(g(bdim) for g in arg_generators),
-                'right': right_args
-            })
-        return out
+    def _get_weights(self, weight: Optional[torch.Tensor]) -> torch.Tensor:
+        if not torch.jit.is_scripting():
+            # If we're not scripting, then we're in Python and `weight` could be a List[Tensor]
+            # deal with that:
+            weight = self._prep_weights_python(weight)
+        if weight is None:
+            if self.weight_numel > 0 and not self.internal_weights:
+                raise RuntimeError("Weights must be provided when the TensorProduct does not have `internal_weights`")
+            return self.weight
+        else:
+            if self.shared_weights:
+                assert weight.shape == (self.weight_numel,), "Invalid weight shape"
+            else:
+                assert weight.shape[-1] == self.weight_numel, "Invalid weight shape"
+                assert weight.ndim > 1, "When shared weights is false, weights must have batch dimension"
+            return weight
 
-    def right(self, features_2, weight=None):
+    @torch.jit.export
+    def right(self, features_2, weight: Optional[torch.Tensor] = None):
         r"""evaluate partially :math:`w x \cdot \otimes y`
 
         It returns an operator in the form of a matrix.
@@ -724,13 +679,11 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
         `torch.Tensor`
             tensor of shape ``(..., irreps_in1.dim, irreps_out.dim)``
         """
-        with torch.autograd.profiler.record_function(repr(self)):
-            weight = self.prepare_weight_list(weight)
-            wigners = [getattr(self, f"C{i}") for i in range(len(self.wigners))]
+        with torch.autograd.profiler.record_function(self._profiling_str):
+            real_weight = self._get_weights(weight)
+            return self._compiled_main_right(features_2, real_weight, self._wigner_buf)
 
-            return self._compiled_main_right(features_2, weight, wigners)
-
-    def forward(self, features_1, features_2, weight=None):
+    def forward(self, features_1, features_2, weight: Optional[torch.Tensor] = None):
         r"""evaluate :math:`w x \otimes y`
 
         Parameters
@@ -753,11 +706,9 @@ class TensorProduct(CodeGenMixin, torch.nn.Module):
         `torch.Tensor`
             tensor of shape ``(..., irreps_out.dim)``
         """
-        with torch.autograd.profiler.record_function(repr(self)):
-            weight = self.prepare_weight_list(weight)
-            wigners = [getattr(self, f"C{i}") for i in range(len(self.wigners))]
-
-            return self._compiled_main_out(features_1, features_2, weight, wigners)
+        with torch.autograd.profiler.record_function(self._profiling_str):
+            real_weight = self._get_weights(weight)
+            return self._compiled_main_out(features_1, features_2, real_weight, self._wigner_buf)
 
     def visualize(self):  # pragma: no cover
         import numpy as np
@@ -895,18 +846,14 @@ class FullyConnectedTensorProduct(TensorProduct):
         irreps_in2 = o3.Irreps(irreps_in2).simplify()
         irreps_out = o3.Irreps(irreps_out).simplify()
 
-        in1 = [(mul, ir, 1.0) for mul, ir in irreps_in1]
-        in2 = [(mul, ir, 1.0) for mul, ir in irreps_in2]
-        out = [(mul, ir, 1.0) for mul, ir in irreps_out]
-
         instr = [
             (i_1, i_2, i_out, 'uvw', True, 1.0)
-            for i_1, (_, (l_1, p_1)) in enumerate(irreps_in1)
-            for i_2, (_, (l_2, p_2)) in enumerate(irreps_in2)
-            for i_out, (_, (l_out, p_out)) in enumerate(irreps_out)
-            if abs(l_1 - l_2) <= l_out <= l_1 + l_2 and p_1 * p_2 == p_out
+            for i_1, (_, ir_1) in enumerate(irreps_in1)
+            for i_2, (_, ir_2) in enumerate(irreps_in2)
+            for i_out, (_, ir_out) in enumerate(irreps_out)
+            if ir_out in ir_1 * ir_2
         ]
-        super().__init__(in1, in2, out, instr, normalization, internal_weights, shared_weights)
+        super().__init__(irreps_in1, irreps_in2, irreps_out, instr, normalization, internal_weights, shared_weights)
 
 
 class ElementwiseTensorProduct(TensorProduct):